Moved rest of border drawing to abstract system
