--- src/layout/core/borders/draw/color.rs
@@ -1,27 +1,0 @@
-/// Color to draw to the screen, including the alpha channel.
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Hash)]
-pub struct Color {
-    red: u8,
-    green: u8,
-    blue: u8,
-    alpha: u8
-}
-
-
-impl Color {
-    /// Makes a new solid color, with no transparency.
-    pub fn solid_color(red: u8, green: u8, blue: u8) -> Self {
-        Color {
-            red: red,
-            green: green,
-            blue: blue,
-            alpha: 255
-        }
-    }
-
-    /// Gets the values of the colors, in this order:
-    /// (Red, Green, Blue, Alpha)
-    pub fn values(&self) -> (u8, u8, u8, u8) {
-        (self.red, self.green, self.blue, self.alpha)
-    }
-}
