--- src/layout/core/borders/draw/mod.rs
@@ -1,8 +1,0 @@
-mod color;
-mod base;
-mod border_draw;
-
-pub use self::color::Color;
-pub use self::base::{BaseDraw, Drawable, DrawErr};
-pub use self::border_draw::BorderDraw;
-
