+++ src/render/mod.rs
@@ -1,7 +1,7 @@
 mod renderable;
-mod base;
+mod draw;
 mod color;
 
 pub use self::renderable::Renderable;
-pub use self::base::BaseDraw;
+pub use self::draw::{Drawable, DrawErr, BaseDraw};
 pub use self::color::Color;
